Move CraftingRecipes#Builder to be exposed to addons
