+++ src/data/java/net/dries007/tfc/data/recipes/CraftingRecipes.java
@@ -74,6 +74,8 @@ import net.dries007.tfc.common.recipes.outputs.ItemStackProvider;
 import net.dries007.tfc.common.recipes.outputs.MealModifier;
 import net.dries007.tfc.util.Metal;
 
+import static net.dries007.tfc.util.DataHelpers.Builder;
+
 public interface CraftingRecipes extends Recipes
 {
     default void craftingRecipes()
@@ -1368,217 +1370,4 @@ public interface CraftingRecipes extends Recipes
         return AndIngredient.of(food, NotRottenIngredient.INSTANCE);
     }
 
-    /**
-     * A recipe builder capable of building shaped, shapeless recipes, optionally with both output and remainder features
-     * of advanced shaped / shapeless recipes. It has some preliminary validations to ensure legal recipes are built
-     */
-    class Builder
-    {
-        final BiConsumer<String, Recipe<?>> onFinish;
-        @Nullable String name = null;
-
-        final List<ItemStackModifier> remainder = new ArrayList<>(); // For advanced recipes, remainder modifiers
-        final List<ItemStackModifier> outputs = new ArrayList<>(); // For advanced recipes, output modifiers
-        final NonNullList<Ingredient> ingredients = NonNullList.create(); // Shapeless recipes only
-        final List<String> pattern = new ArrayList<>(); // Shaped recipes only
-        final ImmutableMap.Builder<Character, Ingredient> keys = ImmutableMap.builder();
-        int inputRow = 0, inputCol = 0;
-        @Nullable Ingredient primaryInput = null;
-        boolean needsAdvInput = false, hasAdvInputShaped = false, hasAdvInputShapeless = false;
-
-        Builder(BiConsumer<String, Recipe<?>> onFinish)
-        {
-            this.onFinish = onFinish;
-        }
-
-        void useTool(TagKey<Item> tool, ItemLike input, ItemLike output)
-        {
-            input(input).inputIsPrimary(tool).damageInputs().shapeless(output);
-        }
-
-        void useTool(TagKey<Item> tool, Ingredient input, ItemLike output, int count)
-        {
-            input(input).inputIsPrimary(tool).damageInputs().shapeless(output, count);
-        }
-
-        void bricksWithMortar(ItemLike brick, ItemLike bricks, int count)
-        {
-            input('Y', TFCItems.MORTAR).input('X', brick).pattern("XYX", "YXY", "XYX").shaped(bricks, count);
-        }
-
-        void to3x3(Ingredient input, ItemLike storage)
-        {
-            input('X', input).pattern("XXX", "XXX", "XXX").shaped(storage);
-        }
-
-        void from3x3(Ingredient input, ItemLike item)
-        {
-            input(input).shapeless(item, 9);
-        }
-
-        void to2x2(ItemLike input, ItemLike output, int count)
-        {
-            input('X', input).pattern("XX", "XX").shaped(output, count);
-        }
-
-        Builder damageInputs()
-        {
-            remainder.add(DamageCraftingRemainderModifier.INSTANCE);
-            return this;
-        }
-
-        Builder copyOldestFood()
-        {
-            outputs.add(CopyOldestFoodModifier.INSTANCE);
-            return this;
-        }
-
-        Builder copyFood()
-        {
-            outputs.add(CopyFoodModifier.INSTANCE);
-            return this;
-        }
-
-        Builder copyForging()
-        {
-            needsAdvInput = true;
-            return addOutputModifier(CopyForgingBonusModifier.INSTANCE);
-        }
-
-        Builder copyInput()
-        {
-            needsAdvInput = true;
-            return addOutputModifier(CopyInputModifier.INSTANCE);
-        }
-
-        Builder addGlass()
-        {
-            needsAdvInput = true;
-            return addOutputModifier(AddGlassModifier.INSTANCE);
-        }
-
-        Builder addPowder() {return addOutputModifier(AddPowderModifier.INSTANCE);}
-
-        Builder addBait() {return addOutputModifier(AddBaitToRodModifier.INSTANCE);}
-
-        Builder extraProduct(ItemLike item) {return extraProduct(item, 1);}
-
-        Builder extraProduct(ItemLike item, int count)
-        {
-            remainder.add(new ExtraProductModifier(new ItemStack(item, count)));
-            return this;
-        }
-
-        Builder addTrait(Holder<FoodTrait> trait) {return addOutputModifier(AddTraitModifier.of(trait));}
-
-        Builder addOutputModifier(ItemStackModifier modifier)
-        {
-            outputs.add(modifier);
-            return this;
-        }
-
-        Builder input(ItemLike item) {return input(item, 1);}
-
-        Builder input(ItemLike item, int count) {return input(Ingredient.of(item), count);}
-
-        Builder input(TagKey<Item> item) {return input(item, 1);}
-
-        Builder input(TagKey<Item> item, int count) {return input(Ingredient.of(item), count);}
-
-        Builder input(Ingredient item) {return input(item, 1);}
-
-        Builder input(Ingredient item, int count)
-        {
-            for (int n = 0; n < count; n++) ingredients.add(item);
-            return this;
-        }
-
-        Builder inputIsPrimary(ItemLike item) {return inputIsPrimary(Ingredient.of(item));}
-
-        Builder inputIsPrimary(TagKey<Item> item) {return inputIsPrimary(Ingredient.of(item));}
-
-        Builder inputIsPrimary(Ingredient item)
-        {
-            primaryInput = item;
-            hasAdvInputShapeless = true;
-            return input(item);
-        }
-
-        Builder input(char key, TagKey<Item> input) {return input(key, Ingredient.of(input));}
-
-        Builder input(char key, ItemLike input) {return input(key, Ingredient.of(input));}
-
-        Builder input(char key, Ingredient input)
-        {
-            keys.put(key, input);
-            return this;
-        }
-
-        Builder source(int row, int col)
-        {
-            inputRow = row;
-            inputCol = col;
-            hasAdvInputShaped = true;
-            return this;
-        }
-
-        Builder pattern(String... pattern)
-        {
-            this.pattern.addAll(List.of(pattern));
-            return this;
-        }
-
-        void shapeless(String name)
-        {
-            this.name = name;
-            shapeless(ItemStack.EMPTY);
-        }
-
-        void shapeless(ItemLike output) {shapeless(output, 1);}
-
-        void shapeless(ItemLike output, int count) {shapeless(new ItemStack(output, count));}
-
-        void shapeless(ItemStack output)
-        {
-            assert pattern.isEmpty() && keys.build().isEmpty() : "Mixing shaped and shapeless recipes";
-            assert hasAdvInputShapeless || !needsAdvInput : "Missing a .inputIsPrimary(Ingredient) for a recipe which depends on input";
-            assert !outputs.isEmpty() || !output.isEmpty() : "Either non-empty output, or output modifiers must be present";
-
-            onFinish.accept(name, isAdvanced()
-                ? new AdvancedShapelessRecipe(ingredients, ItemStackProvider.of(output, outputs), remainder(), Optional.ofNullable(primaryInput))
-                : new ShapelessRecipe("", CraftingBookCategory.MISC, output, ingredients));
-        }
-
-        void shaped(String name)
-        {
-            this.name = name;
-            shaped(ItemStack.EMPTY);
-        }
-
-        void shaped(ItemLike output) {shaped(output, 1);}
-
-        void shaped(ItemLike output, int count) {shaped(new ItemStack(output, count));}
-
-        void shaped(ItemStack output)
-        {
-            assert ingredients.isEmpty() : "Mixing shaped and shapeless recipes";
-            assert hasAdvInputShaped || !needsAdvInput : "Missing a .source(int, int) for a recipe which depends on input";
-            assert !outputs.isEmpty() || !output.isEmpty() : "Either non-empty output, or output modifiers must be present";
-
-            final ShapedRecipePattern pattern = ShapedRecipePattern.of(keys.build(), this.pattern);
-            onFinish.accept(name, isAdvanced()
-                ? new AdvancedShapedRecipe(pattern, true, ItemStackProvider.of(output, outputs), remainder(), inputRow, inputCol)
-                : new ShapedRecipe("", CraftingBookCategory.MISC, pattern, output));
-        }
-
-        private Optional<ItemStackProvider> remainder()
-        {
-            return remainder.isEmpty() ? Optional.empty() : Optional.of(ItemStackProvider.of(ItemStack.EMPTY, remainder));
-        }
-
-        private boolean isAdvanced()
-        {
-            return !remainder.isEmpty() || !outputs.isEmpty();
-        }
-    }
 }
